pyproject.toml
[tool.poetry]
name = "gym"
version = "0.1.0"
description = ""
authors = ["Olga Smotrova <olgahikaritattoo@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.6"
djangorestframework = "^3.14.0"
psycopg2 = "^2.9.5"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
